requirements.txt
click==8.1.3
colorama==0.4.6
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
setuptools==67.6.1
Werkzeug==2.2.3
wheel==0.40.0
zipp==3.15.0
mysql-connector-python==8.0.32
pandas==1.5.3
SQLAlchemy==2.0.8
gunicorn==20.1.0
Flask>=2.2.2
